+++ src/am_map.c
@@ -2649,8 +2649,15 @@ static void AM_drawWalls(void)
   // [Nugget] Tag Finder from PrBoomX
   for (int i = 0;  i < array_size(crossmarks);  i++)
   {
-    AM_drawLineCharacter(cross_mark, NUMCROSSMARKLINES, 128<<MAPBITS, 0,
-                         crossmarks[i].color, crossmarks[i].x, crossmarks[i].y);
+    AM_drawLineCharacter(
+      cross_mark,
+      NUMCROSSMARKLINES,
+      ((int64_t) scale_ftom * 24) * current_video_height / SCREENHEIGHT,
+      0,
+      crossmarks[i].color,
+      crossmarks[i].x,
+      crossmarks[i].y
+    );
   }
   array_clear(crossmarks);
 }